# Conflicts:
#	Dev/AcceptanceTesting.sln
#	Dev/Dev2.Activities/Activities/DsfNativeActivity.cs
#	Dev/Dev2.Instrumentation/Dev2.Instrumentation.csproj
#	Dev/Dev2.Studio.Core/DataList/DataListViewModel.cs
#	Dev/Dev2.Studio/App.xaml.cs
#	Dev/Dev2.Studio/ViewModels/ShellViewModel.cs
#	Dev/Dev2.Studio/ViewModels/WorkSurface/WorkSurfaceContextViewModel.cs
#	Dev/Dev2.Studio/ViewModels/Workflow/WorkflowDesignerViewModel.cs
#	Dev/Dev2.Studio/Views/DataList/DataListView.xaml.cs
#	Dev/Studio.sln
#	Dev/Warewolf.Studio.ViewModels/MenuViewModel.cs

+++ Dev/Dev2.Activities/Activities/DsfExecuteCommandLineActivity.cs
@@ -90,10 +90,7 @@ namespace Dev2.Activities
         }
 
 
-        public override List<string> GetOutputs()
-        {
-            return new List<string> { CommandResult };
-        }
+        public override List<string> GetOutputs() => new List<string> { CommandResult };
         #region Overrides of DsfNativeActivity<string>
 
         public DsfExecuteCommandLineActivity()
@@ -133,42 +130,40 @@ namespace Dev2.Activities
                     AddDebugInputItem(new DebugEvalResult(CommandFileName, "Command", dataObject.Environment, update));
                 }
                 var itr = new WarewolfIterator(dataObject.Environment.Eval(CommandFileName, update));
-                if(!allErrors.HasErrors())
+                if (allErrors.HasErrors())
                 {
-                    var counter = 1;
-                    while (itr.HasMoreData())
+                    return;
+                }
+                var counter = 1;
+                while (itr.HasMoreData())
+                {
+                    var val = itr.GetNextValue();
                     {
-                        var val = itr.GetNextValue();
+                        if(string.IsNullOrEmpty(val))
                         {
-                            if(string.IsNullOrEmpty(val))
-                            {
-                                throw new Exception(ErrorResource.EmptyScript);
-                            }
-                            if (!ExecuteProcess(val, exeToken, out StreamReader errorReader, out StringBuilder outputReader))
-                            {
-                                return;
-                            }
-
-                            allErrors.AddError(errorReader.ReadToEnd());
-                            var bytes = Encoding.Default.GetBytes(outputReader.ToString().Trim());
-                            var readValue = Encoding.ASCII.GetString(bytes).Replace("?", " ");
-
-                            foreach (var region in DataListCleaningUtils.SplitIntoRegions(CommandResult))
-                            {
-                                dataObject.Environment?.Assign(region, readValue, update == 0 ? counter : update);
-                            }
-                            counter++;
-                            errorReader.Close();
+                            throw new Exception(ErrorResource.EmptyScript);
                         }
-                    }
-
-                    if(dataObject.IsDebugMode() && !allErrors.HasErrors())
-                    {
-                        if(!string.IsNullOrEmpty(CommandResult))
+                        if (!ExecuteProcess(val, exeToken, out StreamReader errorReader, out StringBuilder outputReader))
                         {
-                            AddDebugOutputItem(new DebugEvalResult(CommandResult, "", dataObject.Environment, update));
+                            return;
                         }
+
+                        allErrors.AddError(errorReader.ReadToEnd());
+                        var bytes = Encoding.Default.GetBytes(outputReader.ToString().Trim());
+                        var readValue = Encoding.ASCII.GetString(bytes).Replace("?", " ");
+
+                        foreach (var region in DataListCleaningUtils.SplitIntoRegions(CommandResult))
+                        {
+                            dataObject.Environment?.Assign(region, readValue, update == 0 ? counter : update);
+                        }
+                        counter++;
+                        errorReader.Close();
                     }
+                }
+
+                if (dataObject.IsDebugMode() && !allErrors.HasErrors() && !string.IsNullOrEmpty(CommandResult))
+                {
+                    AddDebugOutputItem(new DebugEvalResult(CommandResult, "", dataObject.Environment, update));
                 }
             }
             catch(Exception e)
@@ -209,7 +204,6 @@ namespace Dev2.Activities
                     DispatchDebugState(dataObject, StateType.Before, update);
                     DispatchDebugState(dataObject, StateType.After, update);
                 }
-
             }
         }
 
@@ -239,13 +233,11 @@ namespace Dev2.Activities
                 while (!_process.HasExited && !executionToken.IsUserCanceled)
                 {
                     reader.Append(_process.StandardOutput.ReadToEnd());
-                    if (!_process.HasExited && _process.Threads.Cast<ProcessThread>().Any(a => a.ThreadState == System.Diagnostics.ThreadState.Wait && a.WaitReason == ThreadWaitReason.UserRequest))
+                    if (!_process.HasExited && _process.Threads.Cast<ProcessThread>().Any(a => a.ThreadState == System.Diagnostics.ThreadState.Wait && a.WaitReason == ThreadWaitReason.UserRequest) && !_process.HasExited)
                     {
-                        if (!_process.HasExited)
-                        {
-                            _process.Kill();
-                        }
+                        _process.Kill();
                     }
+
                     if (ModalChecker.IsWaitingForUserInput(_process))
                     {
                         _process.Kill();
@@ -322,44 +314,7 @@ namespace Dev2.Activities
             ProcessStartInfo psi;
             if(val.StartsWith("\""))
             {
-                // we have a quoted string for the cmd portion
-                var idx = val.IndexOf("\" \"", StringComparison.Ordinal);
-                if(idx < 0)
-                {
-                    // account for "xxx" arg
-                    idx = val.IndexOf("\" ", StringComparison.Ordinal);
-                    if(idx < 0)
-                    {
-                        val += Environment.NewLine + "exit";
-                        psi = new ProcessStartInfo("cmd.exe", "/Q /C " + val);
-                    }
-                    else
-                    {
-                        var cmd = val.Substring(0, idx + 1);// keep trailing "
-                        if(File.Exists(cmd))
-                        {
-                            var args = val.Substring(idx + 2);
-                            psi = new ProcessStartInfo("cmd.exe", "/Q /C " + cmd + " " + args);
-                        }
-                        else
-                        {
-                            psi = ExecuteSystemCommand(val);
-                        }
-                    }
-                }
-                else
-                {
-                    var cmd = val.Substring(0, idx + 1);// keep trailing "
-                    if(File.Exists(cmd))
-                    {
-                        var args = val.Substring(idx + 2);
-                        psi = new ProcessStartInfo("cmd.exe", "/Q /C " + cmd + " " + args);
-                    }
-                    else
-                    {
-                        psi = ExecuteSystemCommand(val);
-                    }
-                }
+                psi = ExecuteQuotedExpression(ref val);
             }
             else
             {
@@ -393,6 +348,51 @@ namespace Dev2.Activities
                 psi.RedirectStandardOutput = true;
                 psi.CreateNoWindow = true;
                 psi.Verb = "runas";
+            }
+
+            return psi;
+        }
+
+        private static ProcessStartInfo ExecuteQuotedExpression(ref string val)
+        {
+            ProcessStartInfo psi;
+            // we have a quoted string for the cmd portion
+            var idx = val.IndexOf("\" \"", StringComparison.Ordinal);
+            if (idx < 0)
+            {
+                // account for "xxx" arg
+                idx = val.IndexOf("\" ", StringComparison.Ordinal);
+                if (idx < 0)
+                {
+                    val += Environment.NewLine + "exit";
+                    psi = new ProcessStartInfo("cmd.exe", "/Q /C " + val);
+                }
+                else
+                {
+                    var cmd = val.Substring(0, idx + 1);// keep trailing "
+                    if (File.Exists(cmd))
+                    {
+                        var args = val.Substring(idx + 2);
+                        psi = new ProcessStartInfo("cmd.exe", "/Q /C " + cmd + " " + args);
+                    }
+                    else
+                    {
+                        psi = ExecuteSystemCommand(val);
+                    }
+                }
+            }
+            else
+            {
+                var cmd = val.Substring(0, idx + 1);// keep trailing "
+                if (File.Exists(cmd))
+                {
+                    var args = val.Substring(idx + 2);
+                    psi = new ProcessStartInfo("cmd.exe", "/Q /C " + cmd + " " + args);
+                }
+                else
+                {
+                    psi = ExecuteSystemCommand(val);
+                }
             }
 
             return psi;
@@ -447,7 +447,7 @@ namespace Dev2.Activities
 
         #region Overrides of DsfNativeActivity<string>
 
-        public override List<DebugItem> GetDebugInputs(IExecutionEnvironment dataList, int update)
+        public override List<DebugItem> GetDebugInputs(IExecutionEnvironment env, int update)
         {
             foreach(IDebugItem debugInput in _debugInputs)
             {
@@ -456,7 +456,7 @@ namespace Dev2.Activities
             return _debugInputs;
         }
 
-        public override List<DebugItem> GetDebugOutputs(IExecutionEnvironment dataList, int update)
+        public override List<DebugItem> GetDebugOutputs(IExecutionEnvironment env, int update)
         {
             foreach(IDebugItem debugOutput in _debugOutputs)
             {
@@ -465,15 +465,9 @@ namespace Dev2.Activities
             return _debugOutputs;
         }
 
-        public override IList<DsfForEachItem> GetForEachInputs()
-        {
-            return GetForEachItems(CommandFileName, CommandPriority.ToString());
-        }
+        public override IList<DsfForEachItem> GetForEachInputs() => GetForEachItems(CommandFileName, CommandPriority.ToString());
 
-        public override IList<DsfForEachItem> GetForEachOutputs()
-        {
-            return GetForEachItems(CommandResult);
-        }
+        public override IList<DsfForEachItem> GetForEachOutputs() => GetForEachItems(CommandResult);
 
         public void Dispose()
         {
@@ -486,8 +480,16 @@ namespace Dev2.Activities
 
         public bool Equals(DsfExecuteCommandLineActivity other)
         {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
+            if (ReferenceEquals(null, other))
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+
             return base.Equals(other) 
                 && string.Equals(CommandFileName, other.CommandFileName) 
                 && string.Equals(CommandResult, other.CommandResult)
@@ -496,9 +498,21 @@ namespace Dev2.Activities
 
         public override bool Equals(object obj)
         {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != this.GetType()) return false;
+            if (ReferenceEquals(null, obj))
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, obj))
+            {
+                return true;
+            }
+
+            if (obj.GetType() != this.GetType())
+            {
+                return false;
+            }
+
             return Equals((DsfExecuteCommandLineActivity) obj);
         }
